wallet: Make WalletInitInterface and DummyWalletInit private
