+++ src/qt/bitcoin.cpp
@@ -35,9 +35,6 @@
 #include <util.h>
 #include <warnings.h>
 
-#ifdef ENABLE_WALLET
-#include <wallet/init.h>
-#endif
 #include <walletinitinterface.h>
 
 #include <memory>
@@ -660,11 +657,6 @@ int main(int argc, char *argv[])
     // Start up the payment server early, too, so impatient users that click on
     // bitcoin: links repeatedly have their payment requests routed to this process:
     app.createPaymentServer();
-
-    // Hook up the wallet init interface
-    g_wallet_init_interface.reset(new WalletInit);
-#else
-    g_wallet_init_interface.reset(new DummyWalletInit);
 #endif
 
     /// 9. Main GUI initialization
