+++ src/wallet/rpcdump.cpp
@@ -6,7 +6,6 @@
 #include <key_io.h>
 #include <rpc/safemode.h>
 #include <rpc/server.h>
-#include <wallet/init.h>
 #include <validation.h>
 #include <script/script.h>
 #include <script/standard.h>
