+++ src/wallet/wallet.cpp
@@ -12,7 +12,6 @@
 #include <consensus/consensus.h>
 #include <consensus/validation.h>
 #include <fs.h>
-#include <wallet/init.h>
 #include <key.h>
 #include <key_io.h>
 #include <keystore.h>
